requirements.txt
pandas
scikit-learn
streamlit
altair
kafka-python-ng
python-dotenv
openpyxl
pyarrow